Get indentation at least somewhat sensible
This fixes:

* The indentation of statements in blocks beyond the first one
* Indentation of nested blocks to not grow and grow the deeper you get

And seems somewhat simpler than the code it replaces. One important bit
was changing the grammar to not include leading whitespace inside of
the statement list node.

There'll be more needed for continuation lines, but this at least feels
a heck of a lot more usable than before.