Numerous indentation/formatting fixes
This deals with a lot of the "bogus indent at end of file" issues, as
well as fixing the wrong indentation of `else` when pasting code. It
achieves this by making the grammar a parse whitespace differently, so
there's a good chance it breaks numerous parsing tests. It's not quite
perfect yet; for some reason, even though we return "continuation" at
the end of the file when it really is one correctly now, it doesn't do
the indent properly. It also doesn't quite go right if you hit enter at
the "top level" of the final incomplete statement intending to put some
more stuff on the next line, and you didn't already write an infix on
the previous one or some such. But overall, this feels better.

Also untodo the formatter test, which now passes.