Get find usages working better
The indexing mechanism is a bit Java-specific, unfortunately, and in
places we can't really patch. So, we always index things in a rather
weird way that splits the sigils up the way that the Java indexing
would, so that the hardcoded Java word semantics will then find stuff.

Finally, there's some curious metadata interface that lets us give a
different name to an element, which find usages looks at. This is a
bit weird, but lets us pretend (hopefully without side-effects anywhere
else) that `has $.foo` is really declaring `foo`, so that we then only
look for `foo` elsewhere in the tree; that gets us a text match. The
use of PsiReference resolution then avoids any false positives in the
results.