Get array literal formatting working much better
Alas, this does cost us the hash literal multiline test; it sort of
works half of the time, which is very frustrating, but I've not been
able to figure out why. In the meantime, it's nice to have the array
formatting rather than omit it just because of a corner case for the
hash literals.