Let us auto-complete lower-cased types/constants
We don't know if we're entering a sub name or term name when we are
auto-completing, so offer both. This means we can now complete the
`blob8` when typing something like `blob8.new`, as well as making
suggestions of lower-case constant names. Once we get `my \foo = ...`
contributing its symbol, we'll also be able to auto-complete those too.