More immediate display of Int/Rat/Num/Str/Scalar
Instead of making the user have to dig in to these objects, and for Rat
to even do the numerator/denominator division themselves, we now try to
show a nicer immediate presentation of their content. A Scalar likewise
has its content retrieved and presented as `$ = ...`. Thus, for a Scalar
holding an Int, you now see something like `$ = 42` without having to
open up any child nodes.