Optimize out repeated Mu/Any methods addition
Previously we added them for every parent recursively from the topmost one to the recent one,
which was not wrong in terms of results, but we had to do more work depending on how
long the parent/composition chain was. Avoid this by adding them only in top
parent (one that does not have any parents).
We still do double work in case if there are more than one top parents,
but this fix does not affect any other parts of the system and
we already improve the situation.