Force Log::Timeline server to use ipv4
The bug lied in the fact the IDE could prefer ipv4 and the program ipv6
(or the opposite), so just use 127.0.0.1 explicitly.

Since we launch the program with a env var, we can't toggle between them
while retrying like it's done in the debugger. Just use a single one, if
it proves to be an issue later (no ipv4 stack, for example), we'll
figure out a new/different way.