Remove old ExternalLibrary-related code
We used those classes showing package names
as nodes for a long time and now it is time
to say goodbye to them, as we are adapting
our implementation to use default types
of nodes. We will likely re-instantiate
some parts of this later to get better UI,
but need to cleanup things first.