Export packages from modules
I am not sure about the full details of the
semantics that is used here, but apparently
at least when it comes to modules, the symbols
declared within a module are exported just fine.
With this commit we make jumps from `route`
subroutine call to Cro::HTTP::Router comp unit
file which uses this module-based schema.