Goto related symbol for rule and action method
So we can move between grammars and their action methods easily. For now
we just look for a method or rule in the same module with the matching
name. This may be overly simplistic in a module with many grammars and
actions, and we might want to look for ways to disambiguate. However, in
the common case it rather helps.

One slight problem: for some reason, the class/role navigation related
symbol provider takes a *really* long time to run, so we get a UI freeze
for several seconds. This explains also why I observed sometimes slow to
appear class/role navigation markers in larger projects: it's glacial!
This will need fixing also before we can ship this new feature.