Tweak formatter making it pass tests
Here we do numerous changes related to simplification
of the formatter structure and working principles,
which allow us to successfully move onto the rule-based
formatting schema. This commit passes the existing
test suite with some adjustments, so we can move onward
from here increasing the coverage of configurable settings.

We also temporarily remove `getChildAttributes` implementation
for now which strips us off a proper newline handling, however
we will soon re-instantiate it in a hopefully more intuitive
and robust way.