Add format postprocessor
This is a workaround for a couple of issues caused by
how we parse things. While there might be others, a prominent case
this code addresses is that a `A.b` is parsed as a method call,
but when the call part is wrapped, it becomes an infix application,
thus having zero indentation. Here we try to double format
to workaround and ensure proper formatting.