Address two issues in refactoring
Firstly, do not add a redundant whitespace durning inlining.
We apparently did it in some cases, but the truth is we don't need to do so,
hence just remove the newline created and update bogus tests.

Secondly, the algorithm of finding an anchor for extracting a variable
was bogus: it wrongly skipped the case when the expression extracted was a
statement already, thus the anchor ended up being FILE.
With formatting refactoring the bug was uncovered and resulted in a NPE,
so considering this case now fixes the issue.