First pass at providing a built-in Raku
Doesn't use the existing REPL, but instead has a simple helper script
(that probably gets less simple as we go on). Offers syntax highlights
and auto-completion, including of things that show up in previous lines
(as well as symbols from `use` statements, for example). We provide for
multi-line statements and use Ctrl+Enter to perform execution.