Don't feed highlighter zero-width tokens
These are used to guide the parse, but are useless to the highlighter.
They also confuse various features such as brace insertion, which is the
reason that a paren would sometimes not be inserted at EOF. That wasn't
much of a practical problem until now, but it's really annoying in the
REPL, thus why it's now being fixed.