Migrate project creation means to Comma
Before we were depending on Java components to implement
all this, but with time IntelliJ becomes less and less Java-centric,
so it became a burden too much to bear (in particular, resulting in
build issues).
This commit does first part of this process, providing actions for
project creation and importing. It is based on a simplified version of
what IntelliJ IDEA works with. Next, we will port more components
which will result in lifting restrictions of current Project Structure
implementation as well.