Don't treat coercions like DateTime($foo) as error
They are ambiguous with sub calls, but we can do resolution to avoid
wrongly marking them as an error. The annotation now delegates to the
normal reference resolution of the call to avoid repeating logic. As a
bonus, if we find that we really do have a coercion type, then it will
also apply highlighting to indicate it is a type.