Make symbol cache per project
When two or more projects were opened at once, on documentation
querying an exception occurred. It turns out, for external symbols,
when both projects use the same SDK, external pseudo-PSI cache
was shared and the project who cached the symbols first created
elements with this project. After this, the second project was asking
about the symbols, got them and then gave an exception that current
project and project where the external symbol nodes belong differ.
Fix this case by providing a per-project cache.