Support file cache drop + re-use CORE json
When project A uses sdk S and project B uses sdk S, we cannot use symbols from this sdk
simultaneously, as they are psi nodes and thus tied to a project (and causing exceptions).
So previously we introduced a per-project node cache so that A used symbols A' from S
and B used symbols B' from S. However, it is not necessarily to load e.g. CORE setting
json every time, it is not tied to a project and thus can be re-used over and over.
This fixes performance regression of our test suite that became times slower as it
had to load JSON using slow Raku script over and over for each test.