Only grey out unused variable names
So that in `my $x = 99;` we only highlight the `$x`, not also the
initializer, as we did so far. The highlighting of the initializer is
quite distracting if it's a long expression and you'll be typing the
code using `$x` on the next line, as is normal when writing code.