Cleanup tmp script files some more
We improved the situation with those files recently,
but there was still a hole: threads attempting to get core
setting symbols that were resolving in the meanwhile created
tmp files, but were not removing those, so fix this.
Also take care of exceptions and make sure symbol script in
CORE getting thread is removed as well as docs.