Initial refactor to use Perl6Type
Instead of representing types as strings. This avoids changing a bunch
of things that still need to change, often cheating by just having
things be an unresolved type rather than passing around the resolved
thing. It's enough to get us compiling again. Quite a few things are,
however, already updated as far as they'll need to be.