Tweak variable name completion to focus on variables only
For many years this handler looked at the char typed and if it was something that
can be found in a variable name, but unusual for IntelliJ completion, like `%` or `-`,
manually called for auto-completion popup.
Only recently it became clear this is too broad and we should show the popup when
the token typed is _really_ a variable.