Fix viewing options glitch in profiler UI
Apparently, `ChangeListener` is too broad to do the "React on checkbox click"
thing we intended here, so use `ItemListener` instead.
Plus, take the value to apply to filtering directly from the checkbox each time,
not relying on the shady "Just invert it and hope it is synced on the other end" practice.