Only keep around unparsed JSON for externals
The JSONArray/JSONObject data structures are quite costly in terms of
memory, and contribute to high memory utilization when we have a project
with many dependencies. The main cost we want to avoid is calling out to
Rakudo; the time taken to parse the produced JSON is smaller, so this is
hopefully a good trade-off of reducing memory without a great deal of
slowdown when moving between projects.