Make links in Pod preview try to open an editor
Try and find a file within the project that is the target of the link.
We do this for exact matches, and also implement the simplest bits of
the Raku docs name mangling scheme too (so far as it's reasonable to do
so; it's probably an OK fallback attempt in general). Open up an editor
when we click the link. Non-file:// URLs for now are allowed to just
continue, but we should really open them in the web browser rather than
replacing the Pod preview with them.