Modify the query to include types in new profile, but not in the old profile
Prior to that, for profile comparing, we gathered all the types from the
old one and joined them with the new ones. Thus, types that were introduced
in the new profile were not accounted for, which led to us hiding an important info.

With this we extend the query to obtain the "right join" side of the problem
and then group them properly, so the types existing only in the "new" profile
are returned as well.