Fix odd "re-declaration of trait" errors
Root of the issue was that we detected duplicates in an included
file, e.g. Cro::HTTP::Router, and created warnings for them, which
did not account that the case was happening for the "included" file
and thus ranges are not relevant to the file we're currently
highlighting.

Fix it by simply looking at if the current file belongs to
the right file.