Improve Pod block folding
Paragraph blocks and abbreviated blocks don't come in pairs, but rather
we want to fold consecutive runs of them. Implement that. Also set the
folded text to the first line of the Pod, which seems a better
representation of what has been folded.

Since I can't think what more we'd add, remove claim it's "basic" from
the plugin change log.