Follow advice to fix UI freeze
For whatever reason, there was a UI freeze for a while and it indicated
it's due to the Comma new project action, and the message suggested to
run it on the BGT. Well, OK, I guess we can; the built-in new project
action does that too. Not sure how it could freeze for so long, but
anyway, this seems to help.