Introducing Comma 2.0...
This release represents a major shift for the Comma project in many ways.

From the bottom of my heart, a deep gratitude and thanks to Jonathan Worthington,
Edument, and all past and future contributors to the Comma project. There's been
so much effort put into this codebase and it was an honor to be able to work on it.

The most major change is the shift to the `IntelliJ Platform Gradle Plugin 2.0`. This
allows Comma to be built (as a plugin) without cloning the
`intellij-community` repo and downloading it's entire dependency tree!

This does seem to preclude building Comma as a standalone IDE, at least for the time being.
That appears to be a different beast entirely and we will have to investigate that as
the time and tuits allow.

Other major changes included updating the code to correct for broken and (some) deprecated
API changes, as well as the significant cosmetic adjustment of migrating `Perl6` to `Raku`.
The latter should be almost entirely finished, but there might be some stragglers that I've
missed.

Building should be as simple as adjusting `build.gradle.kts` to ensure that `jetbrainsRuntime`
reflects your OS and CPU architecture, opening this repository in `IntelliJ IDEA`, and selecting
`buildPlugin` from the Gradle build target options.

Next steps:

* Enabling the test suite
* Migrating the repository to the Raku organization on github
* Setting up multi-platform releases through CI/CD
* Setting up plugin signing and distribution via the IntelliJ Platform
* Adding NQP as another supported language
* Converting the entire project to Kotlin?
* <Your wishlists go here!>

Happy hacking! :D