Cleanup most uses of project SDK and fix dying on old project open
The migration hasn't fully happened to solely use the backup SDK
service for storing the Raku SDK, but it's looking a lot better now.

The major fix here is that we should now be able to successfully
open directories that contain old Comma project .iml files.

It's a bit of a hack but there's no real way around it that I could
see for now. And it should hopefully not remain an issue for very
long.