Centralize the isProjectRakudo logic
Since the new version is essentially global across projects, I've
put it in the `CommaProjectUtil` module as a static method.

This has the bonus dimension of reducing the number of non-cached
checks to zero.

But it also means that if you don't open a project named
`rakudo` or a project in a path ending in `/rakudo` **as the first
project in a session**, you won't get the benefit of hiding these
annoying-for-core-work annotations.

Likewise, any non-`rakudo` projects you open afterwards will
be without these annotations.

These scenarios are suboptimal and I will be actively looking
for alternatives.