Better to do the deletion prior to the project open
This seems to work much better -- otherwise we delete new .idea
files along with our old ones.

Note that this is exactly the same behavior that has been required
already when directing users to `New Project from Existing Sources...`,
in that they have to agree to overwrite the old contents of `.idea`
when building the new project.