Migrate project rakudo detection to a service
This fixes the pre-existing issues where the setting would spill
out to other projects (there was also a bug that meant it would
only work properly when the `rakudo` project was opened first).

I hope the implementation / JIT is smart enough to cache the
result of these calls. I'll keep an eye on performance to make
sure that things aren't getting out of hand.