Store the Rakudo Core project status with the element itself
Maybe the JVM would have already been able to optimize this in the
background, but it felt quite suboptimal to fetch a project-level
service from a project fetched from a module fetched from a
PsIElement. These annotators are constantly processing!

So I added a wrapper to a wrapper class where we can safely stash
this and any other details we want about a (implemented) PsiElement.

The calls still go through a static method in CommaProjectUtils
but I reckon the JVM is able to inline that for a certainty.