Fix go-to-declaration / click through from package to source
This should resolve previously experienced issues with the
functionality for going to the source code of a Raku module
that is used in the project code but which does not belong to the
project.

If installed (currently only checked via `zef`), the module is loaded
into a virtual file and presented to the user.

Probably there is an even cleaner way to do this, but prior to this
the search for the module source would be running on the EDT thread,
which has very low tolerance for even relatively short running blocking
processes. Something would then happen during the resolve process that
would cause the entire IDE to freeze.

The hard work is now accomplished in the background by a combination
of services. This means that there is some amount of synchronization
screwups that can happen, but so far it seems more or less robust
"enough".