Fully migrate away from the old SDK approach
We don't want to have any interactions whatsoever with the
ProjectSdksModel: this should be left only for JDKs.

Fixing this necessitated ripping out just about every piece of
the old project creation wizard and associated changed-mildly-or-not-at-all
project structure packages that were imported from the IntelliJ
codebase.

This means some things might still be a bit broken in the project creation
and import story.

However, we now have a lot more freedom to improve those components now
that we are using Kotlin UI DSL and have no ties to the legacy SDK
infrastructure.

But... user's will likely be prompted one last time for an SDK
that they had already previously set.

Oh, yeah, and we have the external libraries working again and
better than ever!