Add 'zef' and dependency management back in
In addition to a notification bar message, users will now
be presented with a blue underline underneath modules that
are in META6.json dependencies and are known to be in the
ecosystem. There is an associated fix that will then
install the dependency.

This also fixes the issue where modules that are specified
as provided in META6.json were being highlighted by the used
module inspection.